App updated to help with download investigations
It's very hard to find out if we have problems with downloads, since we are downloading many files daily.

So I improved the app interface with features to help with th investigations.

Now it is possible to reprocess downloads directly from the interface without use Postman.

We can also group the download logs by name and refdate.
This is useful to clearly see if the task has correctly downloaded the file.